[Android] Move all `jetpack-compose-primitives` components to `jetpack-compose` namespace.
