+++ packages/expo-ui/jetpack-compose-primitives.d.ts
@@ -1,1 +1,1 @@
-export * from './build/jetpack-compose-primitives';
+export * from './build/jetpack-compose';
